fix: reconnect to the gossip network if we have no peers
previously, we only checked our connection to the gossip network when we
got a new state from the coordinator.
this meant that a network disconnection in the Warmup step would not be
repaired until Warmup ended.

to fix this, we simply check our gossip connections every 10s &
re-bootstrap as much as needed if we're offline.